Seperated the derived DailyUpdater classes into their own files
