+++ csharpcore/GildedRose/DailyUpdater.cs
@@ -29,57 +29,3 @@ public abstract class DailyUpdater
         item.Quality = int.Max(item.Quality - byValue, ItemQuality.MinQuality);
     }
 }
-
-public class DailyUpdaterForRegularItems : DailyUpdater
-{
-    public override void UpdateQuality(Item item)
-    {
-        DecreaseQuality(item);
-    }
-}
-
-public class DailyUpdaterForBetterWithAgeItems : DailyUpdater
-{
-    public override void UpdateQuality(Item item)
-    {
-        IncreaseQuality(item);
-    }
-}
-
-
-public class DailyUpdaterForBackstagePassesItems : DailyUpdater
-{
-    public override void UpdateQuality(Item item)
-    {
-        if (item.SellIn > 9)
-        {
-            IncreaseQuality(item);
-        }
-        else if (item.SellIn > 4)
-        {
-            IncreaseQuality(item, 2);
-        }
-        else if (!IsExpired(item))
-        {
-            IncreaseQuality(item, 3);
-        }
-        else //Expired
-        {
-            DecreaseQuality(item, item.Quality);
-        }
-    }
-}
-
-public class DailyUpdaterForLegendaryItems : DailyUpdater
-{
-    public override void UpdateSellIn(Item item)
-    {
-        // Legendary Items don't change over time
-    }
-    public override void UpdateQuality(Item item)
-    {
-        // Legendary Items don't change over time
-    }
-}
-
-
