+++ csharpcore/GildedRose/DailyUpdaterForBetterWithAgeItems.cs
@@ -1,0 +1,9 @@
+namespace GildedRoseKata;
+
+public class DailyUpdaterForBetterWithAgeItems : DailyUpdater
+{
+    public override void UpdateQuality(Item item)
+    {
+        IncreaseQuality(item);
+    }
+}
