+++ csharpcore/GildedRose/DailyUpdaterForLegendaryItems.cs
@@ -1,0 +1,13 @@
+namespace GildedRoseKata;
+
+public class DailyUpdaterForLegendaryItems : DailyUpdater
+{
+    public override void UpdateSellIn(Item item)
+    {
+        // Legendary Items don't change over time
+    }
+    public override void UpdateQuality(Item item)
+    {
+        // Legendary Items don't change over time
+    }
+}
