+++ csharpcore/GildedRose/DailyUpdaterForRegularItems.cs
@@ -1,0 +1,9 @@
+namespace GildedRoseKata;
+
+public class DailyUpdaterForRegularItems : DailyUpdater
+{
+    public override void UpdateQuality(Item item)
+    {
+        DecreaseQuality(item);
+    }
+}
